+++ test/actions/InitAction-test.js
@@ -7,8 +7,11 @@ const userSettingsFolder = path.join('build', 'usersettings')
 const appPath = path.join('build', 'appsettings')
 const sinon = require('sinon')
 const fsEx = require('fs-extra')
+const nock = require('nock')
 
 describe('InitAction', () => {
+  let userSettings
+
   it('should register', () => {
     const commander = {}
     commander.command = sinon.stub().returns(commander)
@@ -26,53 +29,80 @@ describe('InitAction', () => {
 
   beforeEach(() => {
     process.env.USER_PATH = userSettingsFolder
+    process.env.SGCLOUD_DC_ADDRESS = 'http://test.test'
     fsEx.emptyDirSync(userSettingsFolder)
+    userSettings = new UserSettings()
   })
 
   afterEach((done) => {
-    UserSettings.setInstance()
     delete process.env.USER_PATH
+    delete process.env.SGCLOUD_DC_ADDRESS
     fsEx.remove(userSettingsFolder, done)
   })
 
   it('should throw if user not logged in', (done) => {
-    UserSettings.getInstance().getSession().token = null
+    userSettings.setToken(null)
     const init = new InitAction()
-    init.run(null, (err) => {
-      assert.equal(err.message, 'not logged in')
+    try {
+      init.run(null)
+    } catch (e) {
+      assert.equal(e.message, 'not logged in')
       done()
-    })
+    }
   })
 
   it('should throw if application already initialized', (done) => {
-    UserSettings.getInstance().getSession().token = {}
+    userSettings.setToken({})
     const appId = 'foobarTest'
 
     process.env.APP_PATH = appPath
     const appSettings = new AppSettings()
     fsEx.emptyDirSync(path.join(appPath, AppSettings.SETTINGS_FOLDER))
-    appSettings.setId(appId).save().init()
-    AppSettings.setInstance(appSettings)
+    appSettings.setId(appId)
 
     const init = new InitAction()
 
     init.run(null, (err) => {
       assert.equal(err.message, `The current folder is already initialized for application ${appId}`)
       delete process.env.APP_PATH
-      AppSettings.setInstance()
       fsEx.remove(appPath, done)
     })
   })
 
-  it('should create folders, settings file and save appId', (done) => {
-    AppSettings.setInstance()
+  it('should throw an error because getting the application data as validation fails', (done) => {
     process.env.APP_PATH = appPath
-    UserSettings.getInstance().getSession().token = {}
+    userSettings.setToken({})
 
-    new InitAction().run({appId: 'test'}, () => {
-      delete process.env.APP_PATH
-      AppSettings.setInstance()
-      fsEx.remove(appPath, done)
+    const dcMock = nock(process.env.SGCLOUD_DC_ADDRESS)
+    .get(`/applications/test`)
+    .reply(403, {})
+
+    new InitAction().run({appId: 'test'}, (err) => {
+      fsEx.remove(appPath, () => {
+        assert.equal(err.message, 'The application test is not available or permissions are missing (message: Getting application data failed). Please check the application at developer.shopgate.com!')
+        delete process.env.APP_PATH
+        dcMock.done()
+        done()
+      })
+    })
+  })
+
+  it('should create folders, settings file and save appId', (done) => {
+    process.env.APP_PATH = appPath
+    userSettings.setToken({})
+
+    const dcMock = nock(process.env.SGCLOUD_DC_ADDRESS)
+    .get(`/applications/test`)
+    .reply(200, {})
+
+    new InitAction().run({appId: 'test'}, (err) => {
+      fsEx.remove(appPath, (err2) => {
+        assert.ifError(err)
+        assert.ifError(err2)
+        delete process.env.APP_PATH
+        dcMock.done()
+        done()
+      })
     })
   })