+++ test/app/backend/extensionRuntime/StepExecutor-test.js
@@ -157,8 +157,40 @@ describe('StepExecutor', () => {
     }
     executor.execute(input, stepMeta, (err) => {
       assert.ok(err)
-      assert.ok(err.message.endsWith('notThere.js" not found'))
+      assert.ok(err.message.endsWith('notThere.js not found'))
       done()
+    })
+  })
+
+  it('should fail if step file contains a syntax error', (done) => {
+    const input = {foo: 'bar'}
+    const stepMeta = {
+      id: '@foo/bar',
+      path: '@foo/bar/crashing2.js',
+      meta: {appId: 'shop_123'}
+    }
+
+    let exitCalled = false
+    executor.start = (cb) => {
+      assert.equal(executor.childProcess, undefined)
+      assert.ok(exitCalled)
+      done()
+    }
+
+    const onExit = executor.onExit.bind(executor)
+    executor.onExit = (code, signal) => {
+      if (!exitCalled) {
+        assert.equal(code, 1)
+        assert.equal(signal, null)
+      }
+      assert.ok(executor.childProcess)
+      exitCalled = true
+      onExit(code, signal)
+    }
+
+    executor.execute(input, stepMeta, (err) => {
+      assert.ok(err)
+      assert.equal(err.message, 'Cannot read property \'notThere\' of undefined')
     })
   })