+++ test/app/frontend/FrontendSettings-test.js
@@ -4,7 +4,7 @@
  * This source code is licensed under the Apache 2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-
+const fsEx = require('fs-extra')
 const assert = require('assert')
 const FrontendSettings = require('../../../lib/app/frontend/FrontendSettings')
 
@@ -14,6 +14,7 @@ describe('FrontendSettings', () => {
   beforeEach(() => {
     settingsPath = 'foobar'
     settings = new FrontendSettings(settingsPath)
+    fsEx.writeJsonSync(settings.frontendSettingsFile, {})
   })
 
   it('can be instanciated', () => {